// stores/types/BribeABI.ts
export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export type ClaimRewards = ContractEventLog<{
  from: string;
  reward: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;
export type Deposit = ContractEventLog<{
  from: string;
  tokenId: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;
export type NotifyReward = ContractEventLog<{
  from: string;
  reward: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;
export type Withdraw = ContractEventLog<{
  from: string;
  tokenId: string;
  amount: string;
  0: string;
  1: string;
  2: string;
}>;

export interface BribeABI extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): BribeABI;
  clone(): BribeABI;
  methods: {
    DURATION(): NonPayableTransactionObject<string>;

    PRECISION(): NonPayableTransactionObject<string>;

    _deposit(
      amount: number | string | BN,
      tokenId: number | string | BN
    ): NonPayableTransactionObject<void>;

    _ve(): NonPayableTransactionObject<string>;

    _withdraw(
      amount: number | string | BN,
      tokenId: number | string | BN
    ): NonPayableTransactionObject<void>;

    balanceOf(arg0: number | string | BN): NonPayableTransactionObject<string>;

    batchRewardPerToken(
      token: string,
      maxRuns: number | string | BN
    ): NonPayableTransactionObject<void>;

    checkpoints(
      arg0: number | string | BN,
      arg1: number | string | BN
    ): NonPayableTransactionObject<{
      timestamp: string;
      balanceOf: string;
      0: string;
      1: string;
    }>;

    earned(
      token: string,
      tokenId: number | string | BN
    ): NonPayableTransactionObject<string>;

    factory(): NonPayableTransactionObject<string>;

    getPriorBalanceIndex(
      tokenId: number | string | BN,
      timestamp: number | string | BN
    ): NonPayableTransactionObject<string>;

    getPriorRewardPerToken(
      token: string,
      timestamp: number | string | BN
    ): NonPayableTransactionObject<{
      0: string;
      1: string;
    }>;

    getPriorSupplyIndex(
      timestamp: number | string | BN
    ): NonPayableTransactionObject<string>;

    getReward(
      tokenId: number | string | BN,
      tokens: string[]
    ): NonPayableTransactionObject<void>;

    getRewardForOwner(
      tokenId: number | string | BN,
      tokens: string[]
    ): NonPayableTransactionObject<void>;

    isReward(arg0: string): NonPayableTransactionObject<boolean>;

    lastEarn(
      arg0: string,
      arg1: number | string | BN
    ): NonPayableTransactionObject<string>;

    lastTimeRewardApplicable(
      token: string
    ): NonPayableTransactionObject<string>;

    lastUpdateTime(arg0: string): NonPayableTransactionObject<string>;

    left(token: string): NonPayableTransactionObject<string>;

    notifyRewardAmount(
      token: string,
      amount: number | string | BN
    ): NonPayableTransactionObject<void>;

    numCheckpoints(
      arg0: number | string | BN
    ): NonPayableTransactionObject<string>;

    periodFinish(arg0: string): NonPayableTransactionObject<string>;

    rewardPerToken(token: string): NonPayableTransactionObject<string>;

    rewardPerTokenCheckpoints(
      arg0: string,
      arg1: number | string | BN
    ): NonPayableTransactionObject<{
      timestamp: string;
      rewardPerToken: string;
      0: string;
      1: string;
    }>;

    rewardPerTokenNumCheckpoints(
      arg0: string
    ): NonPayableTransactionObject<string>;

    rewardPerTokenStored(arg0: string): NonPayableTransactionObject<string>;

    rewardRate(arg0: string): NonPayableTransactionObject<string>;

    rewards(arg0: number | string | BN): NonPayableTransactionObject<string>;

    rewardsListLength(): NonPayableTransactionObject<string>;

    supplyCheckpoints(arg0: number | string | BN): NonPayableTransactionObject<{
      timestamp: string;
      supply: string;
      0: string;
      1: string;
    }>;

    supplyNumCheckpoints(): NonPayableTransactionObject<string>;

    totalSupply(): NonPayableTransactionObject<string>;

    userRewardPerTokenStored(
      arg0: string,
      arg1: number | string | BN
    ): NonPayableTransactionObject<string>;
  };
  events: {
    ClaimRewards(cb?: Callback<ClaimRewards>): EventEmitter;
    ClaimRewards(
      options?: EventOptions,
      cb?: Callback<ClaimRewards>
    ): EventEmitter;

    Deposit(cb?: Callback<Deposit>): EventEmitter;
    Deposit(options?: EventOptions, cb?: Callback<Deposit>): EventEmitter;

    NotifyReward(cb?: Callback<NotifyReward>): EventEmitter;
    NotifyReward(
      options?: EventOptions,
      cb?: Callback<NotifyReward>
    ): EventEmitter;

    Withdraw(cb?: Callback<Withdraw>): EventEmitter;
    Withdraw(options?: EventOptions, cb?: Callback<Withdraw>): EventEmitter;

    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };

  once(event: "ClaimRewards", cb: Callback<ClaimRewards>): void;
  once(
    event: "ClaimRewards",
    options: EventOptions,
    cb: Callback<ClaimRewards>
  ): void;

  once(event: "Deposit", cb: Callback<Deposit>): void;
  once(event: "Deposit", options: EventOptions, cb: Callback<Deposit>): void;

  once(event: "NotifyReward", cb: Callback<NotifyReward>): void;
  once(
    event: "NotifyReward",
    options: EventOptions,
    cb: Callback<NotifyReward>
  ): void;

  once(event: "Withdraw", cb: Callback<Withdraw>): void;
  once(event: "Withdraw", options: EventOptions, cb: Callback<Withdraw>): void;
}